Convert bmi_model members from shared_ptr to unique_ptr
This avoids giving the impression that there should ever actually be
shared ownership - a model instance is always exclusively held by the
single adapter instance that created it.

Leave the C++ adapter alone for now due to the more complicated
lifetime logic that it implements.

Leave the Python adapted alone because the unit tests do silly things
with reaching in and grabbing the underlying object.